Add files via upload
The most reliable approach to anticipate the future is to endeavor to comprehend the current, thus we have established our prior target as an examination of the Nifty 500's current condition in order to fully understand and try to develop a positive future opportunity for investment. We have used different index market value based on these index market value we have tried to develop an accurate model that forecast our price more accurately. In this analysis, we use an application of ARIMA to forecast & GARCH model for the prediction of stability based on which we predict the future value & volatility of the indices. To fit the model, we use a cross-validation technique with the observed data of S&P, Nikkie, Ftse, and Nifty 500.